+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Teamup Rooms Changelog
 
+## [Quick Booking] - 2022-06-10
+
+- Polished up the list of rooms with icons and colors that reflect Teamup's colors
+- Support booking rooms through a Quick Add action. Quick Add uses the event length in the search bar dropdown.
+- Added a default title for new events in preferences. This is used for booking romos
+
 ## [Initial Version] - 2022-06-09